fix: config
